==> 1-sentiment_comments.ipynb <==
!pip install -U textblob-de==0.4.3

!python -m textblob.download_corpora

#### Sentiment Analysis Klinikum.de

import pandas as pd
import textblob
from textblob_de import TextBlobDE

df = pd.read_csv('../csv/merged_all.csv')
df.reset_index(drop=True, inplace=True)
df_temp = df.dropna(subset=['Textuelle_Bewertung'], axis=0)

df.shape

df_temp.shape

dikt={}
for i in range(1,25):
    sent=[]
    if df.loc[df['Klinik_ID']==i]['Textuelle_Bewertung'] is not str('') and df.loc[df['Klinik_ID']==i]['Textuelle_Bewertung'] is not None:
        for a in df.loc[df['Klinik_ID']==i]['Textuelle_Bewertung']:
            blob = TextBlobDE(str(a))
            x=blob.sentiment[0]
            sent.append(x)
    dikt[i] = sent

#### Find Number of Comments as Positive Negative and Neutral According to Sentiment Analysis

perc_dic = {}
for key,values in dikt.items():
    positive = 0
    negative = 0
    neutral = 0
    for i in values:
        if i>0:
            positive+=1
        elif i<0:
            negative+=1
        else:
            neutral+=1
    perc_dic[key]= [negative,neutral,positive]

perc_dic

df_ff = pd.DataFrame.from_dict(perc_dic, orient='index', columns=['Negative', 'Neutral', 'Positive'])

df_ff.reset_index(inplace=True)
df_ff = df_ff.rename(columns = {'index':'Klinik_ID'})

df_ff['Total_Comments'] = df_ff['Negative'] + df_ff['Neutral'] + df_ff['Positive']

df_ff

df_ff.to_json(r'number_of_comments_klinikumbew.json')

#### Find All Sentiment Points

new_dic = {}
for key,values in dikt.items():
    a = [(i * 3)+3 for i in values]
    a =sum(a)/len(a)
    new_dic[key]= a

for key in  new_dic.items():
    print(key)

def make_df_from_dic(dic, column_name):
    return pd.DataFrame(list(dic.items()),columns = ['Klinik_ID',column_name]) 

df_kb = make_df_from_dic(new_dic,"KB_Sentiment")

df_kb.to_json(r'klinikbewertung_sentiment.json')

#### Sentiment Analysis GoogleMaps

df_google = pd.read_csv('../csv/merged_all_google.csv')
df_google.reset_index(drop=True, inplace=True)
df_temp_google = df_google.dropna(subset=['Textuelle_Bewertung'], axis=0)

df_google.shape

df_temp_google.shape

dikt={}
for i in df_google['Klinik_ID'].unique():
    sent=[]
    if df_google.loc[df_google['Klinik_ID']==i]['Textuelle_Bewertung'] is not str(''):
        for a in df_google.loc[df_google['Klinik_ID']==i]['Textuelle_Bewertung']:
            blob = TextBlobDE(str(a))
            x=blob.sentiment[0]
            sent.append(x)
    dikt[i]=sent

perc_dic_g = {}

for key,values in dikt.items():
    positive = 0
    negative = 0
    neutral = 0
    for i in values:
        if i>0:
            positive+=1
        elif i<0:
            negative+=1
        else:
            neutral+=1
    perc_dic_g[key]= [negative,neutral,positive]

df_gg = pd.DataFrame.from_dict(perc_dic_g, orient='index', columns=['Negative', 'Neutral', 'Positive'])
df_gg.reset_index(inplace=True)
df_gg = df_gg.rename(columns = {'index':'Klinik_ID'})
df_gg=df_gg.sort_values(by='Klinik_ID', ascending=True)
df_gg.reset_index(drop= True, inplace=True)

df_gg['Total_Comments'] = df_gg['Negative'] + df_gg['Neutral'] + df_gg['Positive']

df_gg

df_gg.to_json(r'number_of_comments_google.json')

new_dic = {}
for key,values in dikt.items():
    a = [(i * 2)+3 for i in values]
    a =sum(a)/len(a)
    new_dic[key]= a

for key in new_dic.items():
    print(key)

#### Gesamtsterne Mean of GoogleMaps

mean_all_comments = {}
for i in df_google['Klinik_ID'].unique():
    a=0
    count=0
    for k in df_google.loc[df_google['Klinik_ID']==i]['Sternebewertung']:
        a=a+int(k.strip(" Stern"))
        count=count+1
        mean = a/count
    mean_all_comments[i] = mean
for i in mean_all_comments.items():
    print(i)

mean_all_comments = []
for i in df_google['Sternebewertung']:
    x = int(i.strip(" Stern"))
    mean_all_comments.append(x)
df_google['Sternebewertung'] = mean_all_comments

df_google.isna().sum()

df_google['Likes'] = df_google['Likes'].fillna(0)

df_google.isna().sum()

df_google.to_csv( "../csv/merged_all_google_updated.csv", encoding="utf-8", index=False)

df_google.to_json(r'merged_all_google_updated.json')

#### mean_only_with_comments 

mean_only_with_comments = {}
for i in df_temp_google['Klinik_ID'].unique():
    a=0
    count=0
    for k in df_temp_google.loc[df_temp_google['Klinik_ID']==i]['Sternebewertung']:
        a=a+int(k.strip(" Stern"))
        count=count+1
        mean = a/count
    mean_only_with_comments[i] = mean
for i in mean_only_with_comments.items():
    print(i)

#### Merge All Gesamtsterne Means in a Dataframe

def make_df_from_dic(dic, column_name):
    return pd.DataFrame(list(dic.items()),columns = ['Klinik_ID',column_name]) 

df_different = make_df_from_dic(new_dic,"Sentiment")

df_mean_withnull = make_df_from_dic(mean_all_comments,"GesamtsterneWithNull")

df_mean_withoutnull = make_df_from_dic(mean_only_with_comments,"GesamtsterneWithoutNull")

df_merge_2 = pd.concat([df_different.set_index('Klinik_ID'),df_mean_withnull.set_index('Klinik_ID'), df_mean_withoutnull.set_index('Klinik_ID')], axis=1, join='inner').sort_values('Klinik_ID', ascending = True).reset_index()

df_merge = df_different.merge(df_mean_withnull, on='Klinik_ID', how='inner')

result_list = []
for x,y in zip(df_merge_2["Sentiment"], df_merge_2["GesamtsterneWithoutNull"]):
    result = x/y
    if result >1:
        result = 1/result   
    result_list.append(((result)*100))
df_merge_2["Accuracy"] =  result_list   

df_merge_2["Sentiment"] = df_merge_2["Sentiment"].round(2)
df_merge_2["GesamtsterneWithNull"] = df_merge_2["GesamtsterneWithNull"].round(2)
df_merge_2["GesamtsterneWithoutNull"] = df_merge_2["GesamtsterneWithoutNull"].round(2)
df_merge_2["Accuracy"] = df_merge_2["Accuracy"].round(2)
df_merge_2

#### All Analysis from Df to Json to be used in WEB

df_merge_2.to_json(r'sentiment_comments_google.json')
