==> 2-linear_reg_klinikum.ipynb <==
#### Fill Null Values in GEstaltung Sterne with Linear Regression

import pandas as pd
import matplotlib.pyplot as plt
import pandas as pd
import pylab as pl
import numpy as np
import math
%matplotlib inline

df = pd.read_csv('../csv/merged_all.csv')
df.reset_index(drop=True, inplace=True)

df.shape

df.info()

df.isna().sum()

#### Delete Null Rows of Text_Bewertung and Gestaltung

df_new = df.dropna(subset=['Textuelle_Bewertung'], axis=0)

df_new = df.dropna(subset=['Gestaltung'], axis=0)

df_new.isna().sum()

df_new.head()

#### Fill Null Values in Gestaltung Sterne

msk = np.random.rand(len(df_new)) < 0.8
train = df_new[msk]
test = df_new[~msk]

<h2 id="multiple_regression_model">Multiple Regression Model</h2>


from sklearn import linear_model
regr = linear_model.LinearRegression()
x = np.asanyarray(train[['Verwaltung','Behandlung','Qualitaet']])
y = np.asanyarray(train[['Gestaltung']])
regr.fit (x, y)
# The coefficients
[coef] = regr.coef_.tolist()
[intercept] = regr.intercept_.tolist()
print ('Coefficients: ', regr.coef_)
print('Intercept: ', regr.intercept_)

<h2 id="prediction">Prediction</h2>


y_hat= regr.predict(test[['Verwaltung','Behandlung','Qualitaet']])
x = np.asanyarray(test[['Verwaltung','Behandlung','Qualitaet']])
y = np.asanyarray(test[['Gestaltung']])
print("Residual sum of squares: %.2f"
      % np.mean((y_hat - y) ** 2))

# Explained variance score: 1 is perfect prediction
print('Variance score: %.2f' % regr.score(x, y))

**explained variance regression score:**  
If $\\hat{y}$ is the estimated target output, y the corresponding (correct) target output, and Var is Variance, the square of the standard deviation, then the explained variance is estimated as follow:

$\\texttt{explainedVariance}(y, \\hat{y}) = 1 - \\frac{Var{ y - \\hat{y}}}{Var{y}}$  
The best possible score is 1.0, lower values are worse.


df.isna().sum()

count = 0
for i,a,b,c in zip(df['Gestaltung'], df['Verwaltung'], df['Behandlung'],df['Qualitaet']):
    if math.isnan(i):
        new_val = intercept + coef[0]*a + coef[1]*b + coef[2]*c
        df.at[count, 'Gestaltung'] = new_val
        #df.iloc[count]['Gestaltung'] = new_val 
        #df.iloc[count, df.columns.get_loc('Gestaltung')] = new_val
        count +=1
    else:
        count += 1

df.isna().sum()

df["Gestaltung"] = df["Gestaltung"].round(2)

df.to_csv( "../csv/merged_all_kb_updated.csv", encoding="utf-8", index=False)

df.to_json(r'merged_all_kb_updated.json')

df.tail(10)

#### From Teilsterne Prediction of the Gesamtsterne

# write your code here
# It has worse accuracy
regr2 = linear_model.LinearRegression()
x = np.asanyarray(train[['Qualitaet','Behandlung','Verwaltung', 'Gestaltung']])
y = np.asanyarray(train[['Gesamte_Sterne']])
regr2.fit (x, y)
# The coefficients
print ('Coefficients: ', regr2.coef_)
print('Intercept: ', regr2.intercept_)
[coef] = regr2.coef_.tolist()
[intercept] = regr2.intercept_.tolist()

y_hat= regr2.predict(test[['Qualitaet','Behandlung','Verwaltung', 'Gestaltung']])
x = np.asanyarray(test[['Qualitaet','Behandlung','Verwaltung', 'Gestaltung']])
y = np.asanyarray(test[['Gesamte_Sterne']])
print("Residual sum of squares: %.2f"
      % np.mean((y_hat - y) ** 2))

# Explained variance score: 1 is perfect prediction
print('Variance score: %.2f' % regr2.score(x, y))

df.isna().sum()

count = 0
for i,a,b,c,d in zip(df['Gesamte_Sterne'], df['Qualitaet'], df['Behandlung'], df['Verwaltung'],df['Gestaltung']):
    new_val = intercept + coef[0]*a + coef[1]*b + coef[2]*c + coef[3]*d
    if(new_val >= 6):
        new_val = 6
    df.at[count, 'Gesamte_Sterne_Predict'] = new_val
    #df.iloc[count]['Gestaltung'] = new_val 
    #df.iloc[count, df.columns.get_loc('Gestaltung')] = new_val
    count +=1

df.isna().sum()

df["Gesamte_Sterne_Predict"] = df["Gesamte_Sterne_Predict"].round(2)

df.head()

## Calculation of Gesamte Sterne for Each Hospitals

#### Values in Klinikumbewertung.de

dikt_klinik={}
for i in df['Klinik_ID'].unique():
    sent=[]
    for j in df.loc[df['Klinik_ID']==i]['Gesamte_Sterne']:
        sent.append(j)
    dikt_klinik[i]=sum(sent)/len(sent)

for key in dikt.items():
    print(key)

#### Our Prediction with Multiple Linear Regression

dikt_prediction={}
for i in df['Klinik_ID'].unique():
    sent=[]
    for j in df.loc[df['Klinik_ID']==i]['Gesamte_Sterne_Predict']:
        sent.append(j)
    dikt_prediction[i]=sum(sent)/len(sent)

for key in dikt.items():
    print(key)

#### Merge All Gesamtsterne Means in a Dataframe

def make_df_from_dic(dic, column_name):
    return pd.DataFrame(list(dic.items()),columns = ['Klinik_ID',column_name]) 

df_prediction = make_df_from_dic(dikt_prediction,"Prediction")

df_klinikum = make_df_from_dic(dikt_klinik,"Klinikumbew_Gesamt")

df_merge = pd.concat([df_prediction.set_index('Klinik_ID'),df_klinikum.set_index('Klinik_ID')], axis=1, join='inner').sort_values('Klinik_ID', ascending = True).reset_index()

df_merge = df_different.merge(df_mean_withnull, on='Klinik_ID', how='inner')

result_list = []
for x,y in zip(df_merge["Prediction"], df_merge["Klinikumbew_Gesamt"]):
    result = x/y
    if result >1:
        result = 1/result   
    result_list.append(((result)*100))
df_merge["Accuracy"] =  result_list   

df_merge["Prediction"] = df_merge["Prediction"].round(2)
df_merge["Klinikumbew_Gesamt"] = df_merge["Klinikumbew_Gesamt"].round(2)
df_merge["Accuracy"] = df_merge["Accuracy"].round(2)
df_merge

#### All Analysis from Df to Json to be used in WEB

df_merge.to_json(r'linear_regression_klinikumbewertungen.json')
