==> 3-create_score.ipynb <==
#### Import Json FIle to DF

pip install yellowbrick

import pandas as pd
from bs4 import BeautifulSoup
from selenium import webdriver
from selenium.webdriver.common.keys import Keys
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
import time
import pandas as pd
import numpy as np
import os, glob
import os.path
from sklearn.cluster import KMeans 

df_lr = pd.read_json (r'linear_regression_klinikumbewertungen.json')
df_google_sent = pd.read_json (r'sentiment_comments_google.json')
df_kb_sent = pd.read_json (r'klinikbewertung_sentiment.json')
df_kb_comments_nr = pd.read_json (r'number_of_comments_klinikumbew.json')
df_google_comments_nr = pd.read_json (r'number_of_comments_google.json')

kliniklist = pd.read_excel(r'..\..\Klinikliste.xlsx')
df_kliniklist = pd.DataFrame(kliniklist, columns= ['Klinikname','Link Google Maps','Link Klinikbewertungen'])
df_kliniklist['klinik_ID'] = [i for i in range(1,25)]
df_kliniklist.reset_index(drop=True, inplace=True)

df_lr.head(1)

df_google_sent.head(1)

df_google_comments_nr.head(1)

df_kb_comments_nr.head(1)

df_kb_sent["KB_Sentiment"] = df_kb_sent["KB_Sentiment"].round(2)

df_merge_kb = pd.concat([df_kb_sent.set_index('Klinik_ID'), df_lr.set_index('Klinik_ID'),
                      df_kb_comments_nr.set_index('Klinik_ID')], axis=1, join='inner').sort_values('Klinik_ID', ascending = True).reset_index()

df_merge_kb.head(2)

df_merge_google = pd.concat([df_google_sent.set_index('Klinik_ID'), df_google_comments_nr.set_index('Klinik_ID')], 
                            axis=1, join='inner').sort_values('Klinik_ID', ascending = True).reset_index()

df_merge_google.head(2)

#### Calculation of a SCHIFA SCORE

def schifa1(df1, c1, c2, c3, c4):
    calculate_point = ((df1[c1]*(2/3)+1) + (df1[c2]*(2/3)+1) + (df1[c3]*(2/3)+1)) * df1[c4]
    return calculate_point.round(2)

df_merge_kb['Total_Kb'] = schifa1(df_merge_kb, "KB_Sentiment", "Prediction", "Klinikumbew_Gesamt", "Total_Comments")

df_merge_kb.head(2)

def schifa2(df1, c1, c2, c3):
    calculate_point = (df1[c1] + df1[c2]) * df1[c3]
    return calculate_point.round(2)

df_merge_google['Total_Kb'] = schifa2(df_merge_google, "Sentiment", "GesamtsterneWithNull", "Total_Comments")

df_merge_google.head(2)

df_outer = pd.merge(df_merge_kb,df_merge_google,on='Klinik_ID',how='outer')

df_outer = df_outer.fillna(0)

df_outer.head(5)

def schifa(df1, c1, c2, c3 , c4):
    calculate_point = (df1[c1] + df1[c2])/ (df1[c3]*3+df1[c4]*2)
    return calculate_point.round(2)

df_outer['SCHIFA'] = schifa(df_outer, "Total_Kb_x", "Total_Kb_y" , "Total_Comments_x", "Total_Comments_y")

#### Grouping

df_outer.loc[df_outer.SCHIFA>2.4, 'Shifa_Score']='Mangelhaft'
df_outer.loc[df_outer.SCHIFA>3.1, 'Shifa_Score']='Ausreichend'
df_outer.loc[df_outer.SCHIFA>3.4, 'Shifa_Score']='Gut'
df_outer.loc[df_outer.SCHIFA>3.6, 'Shifa_Score']='Sehr Gut'
df_outer.loc[df_outer.SCHIFA>3.9, 'Shifa_Score']='Hervorragend'

df_outer.head(5)

df_outer.to_json(r'SCHIFA.json')
